Add support to run scripts on ROCm
The model itself is already compatible with ROCm using the ROCm/DGL
port. These changes, extracted from
https://github.com/ROCm/DeepLearningExamples/pull/3, add support for
ROCm to the example training and inference scripts.

Note that this removes pynvml from the requirements.txt and unlike that
PR, does not add amdsmi. This means that in order to run the scripts,
the user will need to install the appropriate platform-specific GPU
management library. Given that the core DGL dependency is *also* not
included in the requirements.txt, this seems not so terrible, and it
should enable pip installing directly from this git repo without issue
on either platform. I think the right way to do this would be optional
components in setup.py or the new variant wheels, but not fiddling with
that right now.

Co-authored-by: jamesETsmith <james.smith9113@gmail.com>
Co-authored-by: mukh1l <mmallaiy@amd.com>
Co-authored-by: awelling2801 <anuya.welling@amd.com>
Co-authored-by: Vicky Tsang <vicky.tsang@amd.com>