Remove torch.jit.script
torch.jit.script blocks cause graph breaks under torch.compile and
mixing two different compilation options seems a bit like asking for
trouble. It would be possible to write a decorator that lazily invokes
torch.jit.script (which itself is lazy) if not under torch.compile, but
that would be pretty complicated. torch.jit.script is deprecated and the
point of this work is to show improvements with torch.compile, so we'll
take the performance hit in the non-compiled case.